+++ core.h
@@ -1,5 +1,5 @@
-#ifndef ANNEALING_CORE_H
-#define ANNEALING_CORE_H
+#ifndef CORE_H
+#define CORE_H
 
 // Containers & Data types
 #include <string>
@@ -18,6 +18,26 @@
 #define EPSILON 1E-12
 
 std::pair<int,std::vector<int> > globalOptimum;
+
+std::vector<int> initialGreedy(int nodes, const std::vector<std::pair<int,int> > &edge, const std::vector<std::pair<int,int> > &distance){
+    
+    std::vector<int> newConfig(nodes);
+    std::vector<bool> visited(nodes,false);
+    int pos = 1;
+    
+    for(int i=0; i<distance.size(); i++){
+
+        // From edge e, starting from u and going to v
+        int e = distance[i].second;
+        int u = edge[e].first;
+        int v = edge[e].second;
+
+        if(!visited[u]){ newConfig[u] = pos++; visited[u] = true; }
+        if(!visited[v]){ newConfig[v] = pos++; visited[v] = true; }
+    }
+    
+    return newConfig;
+}
 
 int cutValue(const std::vector< std::pair<int,int> > &edge, const std::vector<int> &f, int nodes) {
 
@@ -40,106 +60,6 @@ int cutValue(const std::vector< std::pair<int,int> > &edge, const std::vector<in
     for(int i=0; i<cutwidth.size(); i++) maxCut = cutwidth[i] > maxCut ? cutwidth[i] : maxCut;
     
     return maxCut;
-}
-
-std::vector<int> nextNeighbor(const std::vector <std::pair<int,int> > &edge, std::vector <int>&f, int nodes, int T) { 
-    
-    std::vector <int> newConfig = f;
-
-    // Dinamically controlling neighborhood, making fewers disturbances in lower temperatures
-    double tempFactor = T/(nodes*nodes*4.0); 
-    int neighborhoodSize = (int) ceil(tempFactor*nodes) + 2;
-
-    for(int i=0; i<neighborhoodSize/2; i++){
-        
-        int a = rand() % (nodes-1);
-        int b = rand() % (nodes-1);
-        
-        std::swap(newConfig[a], newConfig[b]);
-    }
-
-    return newConfig;
-}
-
-std::vector<int> initialGreedy(int nodes, const std::vector<std::pair<int,int> > &edge, const std::vector<std::pair<int,int> > &distance){
-    
-    std::vector<int> newConfig(nodes);
-    std::vector<bool> visited(nodes,false);
-    int pos = 1;
-    
-    for(int i=0; i<distance.size(); i++){
-
-        // From edge e, starting from u and going to v
-        int e = distance[i].second;
-        int u = edge[e].first;
-        int v = edge[e].second;
-
-        if(!visited[u]){ newConfig[u] = pos++; visited[u] = true; }
-        if(!visited[v]){ newConfig[v] = pos++; visited[v] = true; }
-    }
-    
-    return newConfig;
-}
-
-void simulatedAnnealing(const std::vector <std::pair<int, int> > &edge, std::vector<int> &f, int nodes) {
-    
-    // Setting initial parameters
-    double T = nodes;
-    double Tmin = 0.001;
-    
-    std::vector <int> solution = f;
-    int best = cutValue(edge,f,nodes);
-    
-    while(T>Tmin){
-    
-        double trials = 0.0;
-        double changes = 0.0;
-        double equilibrium = 0.0;
-        
-        while(trials < SA_MAX_TRIALS) {
-                
-            trials++;
-
-            solution = nextNeighbor(edge, f, nodes, T);
-            int cost = cutValue(edge, solution, nodes);
-            int delta = cost - best;
-            
-            if(cost < globalOptimum.first) {
-                
-                globalOptimum.first = cost;
-                globalOptimum.second = solution;
-            
-            }
-
-            if(delta <= EPSILON) {
-                
-                changes++;
-                best = cost;
-                f = solution;
-            
-            }
-            else {
-                // Accepts a bad move with a given probability
-        
-                double e = 1.0/exp(delta/T);
-                double R = (rand()%100 + 1)/100.0;
-                if(e>R)	{
-
-                    changes++;
-                    best = cost;
-                    f = solution; 
-
-                }
-
-            }
-        }
-        
-        // Cooling schedule
-        equilibrium = (changes*1.0)/(trials*1.0);
-        double k = ( (equilibrium>1)?equilibrium/100.0:equilibrium ) * 0.25;
-        T *= 0.60 + k; 
-
-    } 
 }
 
 std::vector<std::pair<int,int> > getDistances(const std::vector<std::pair<int,int> > &edge, std::vector<int> &f){
